Extend dataset api tests to cover more scenarios
All tests cases for dataset API now involve calling mongo to check
created or updated resources instead of relying on the response to a
request. This not only makes it clear that fields have been
updated/created in our datastore but also due to manipulation by the API
depending on user and request the document may appear different in
datastore; also updates can result in other resources being updated, so
these need to be checked to see whether they have been updated
correctly.